chore: updated tests for download.py
